requirements.txt
pydantic>=2.6,<3
python-dotenv>=1.0,<2
openai>=1.40,<2
vosk>=0.3.45,<0.4
sounddevice>=0.4.7,<0.5
numpy>=1.26,<3
pyttsx3>=2.90,<3
pytest>=8.0,<9
